+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GoogleSubscriberInterpreter.scala
@@ -13,9 +13,11 @@ import com.google.common.util.concurrent.MoreExecutors
 import com.google.pubsub.v1.{PubsubMessage, _}
 import fs2.Stream
 import fs2.concurrent.Queue
-import io.chrisdavenport.log4cats.Logger
+import io.chrisdavenport.log4cats.{Logger, StructuredLogger}
 import io.circe.Decoder
 import io.circe.parser._
+import org.broadinstitute.dsde.workbench.model.TraceId
+import com.google.protobuf.Timestamp
 
 import scala.concurrent.duration.FiniteDuration
 
@@ -65,25 +67,60 @@ object GoogleSubscriberInterpreter {
                                                   queue: fs2.concurrent.Queue[F, Event[MessageType]]
                                                 ): GoogleSubscriberInterpreter[F, MessageType] = new GoogleSubscriberInterpreter[F, MessageType](subscriber, queue)
 
-  def receiver[F[_]: Effect: Logger, MessageType: Decoder](queue: fs2.concurrent.Queue[F, Event[MessageType]]): MessageReceiver = new MessageReceiver() {
+  private[google2] def receiver[F[_]: Effect, MessageType: Decoder](queue: fs2.concurrent.Queue[F, Event[MessageType]])(implicit logger: StructuredLogger[F]): MessageReceiver = new MessageReceiver() {
     override def receiveMessage(message: PubsubMessage, consumer: AckReplyConsumer): Unit = {
+      val parseEvent = for {
+        isJson <- Effect[F].fromEither(parse(message.getData.toStringUtf8)).attempt
+        msg <- isJson match {
+          case Left(_) =>
+            Effect[F].raiseError[MessageType](new Exception(s"${message.getData.toStringUtf8} is not a valid Json"))
+          case Right(json) =>
+            Effect[F].fromEither(json.as[MessageType])
+        }
+        traceId = Option(message.getAttributesMap.get("traceId")).map(s => TraceId(s))
+      } yield Event(msg, traceId, message.getPublishTime, consumer)
+
+      val loggingCtx = Map(
+        "traceId" -> Option(message.getAttributesMap.get("traceId")).getOrElse(""),
+        "publishTime" -> message.getPublishTime.toString
+      )
+
       val result = for {
-        json <- parse(message.getData.toStringUtf8)
-        message <- json.as[MessageType]
-        _ <- queue.enqueue1(Event(message, consumer)).attempt.toIO.unsafeRunSync()
+        res <- parseEvent.attempt
+        _ <- res match {
+          case Right(event) =>
+            for {
+              r <- queue.enqueue1(event).attempt
+
+              _ <- r match {
+                case Left(e) =>
+                  logger.info(loggingCtx)(s"Subscriber fail to enqueue $message due to $e") >> Effect[F].delay(consumer.nack()) //pubsub will resend the message up to ackDeadlineSeconds (this is configed during subscription creation)
+                case Right(_) =>
+                  logger.info(loggingCtx)(s"Subscriber Successfully received $message.")
+              }
+            } yield ()
+          case Left(e) =>
+            logger.info(s"Subscriber fail to decode message ${message} due to ${e}. Going to ack the message") >> Effect[F].delay(consumer.ack())
+        }
       } yield ()
 
-      result match {
-        case Left(e) =>
-          Logger[F].info(s"failed to publish $message to internal Queue due to $e")
-          consumer.nack() //pubsub will resend the message up to ackDeadlineSeconds (this is configed during subscription creation
-        case Right(_) =>
-          ()
-      }
+      result.toIO.unsafeRunSync
     }
   }
 
-  def subscriber[F[_]: Effect: Logger, MessageType: Decoder](subscriberConfig: SubscriberConfig, queue: fs2.concurrent.Queue[F, Event[MessageType]]): Resource[F, Subscriber] = {
+  private[google2] def stringReceiver[F[_]: Effect](queue: fs2.concurrent.Queue[F, Event[String]]): MessageReceiver = new MessageReceiver() {
+    override def receiveMessage(message: PubsubMessage, consumer: AckReplyConsumer): Unit = {
+      val enqueueAction = queue.enqueue1(
+        Event(message.getData.toStringUtf8,
+        Option(message.getAttributesMap.get("traceId")).map(s => TraceId(s)),
+        message.getPublishTime,
+        consumer
+      ))
+      enqueueAction.toIO.unsafeRunSync
+    }
+  }
+
+  def subscriber[F[_]: Effect: StructuredLogger, MessageType: Decoder](subscriberConfig: SubscriberConfig, queue: fs2.concurrent.Queue[F, Event[MessageType]]): Resource[F, Subscriber] = {
     val subscription = ProjectSubscriptionName.of(subscriberConfig.projectTopicName.getProject, subscriberConfig.projectTopicName.getTopic)
 
     for {
@@ -101,7 +138,25 @@ object GoogleSubscriberInterpreter {
     } yield sub
   }
 
-  private def subscriberResource[MessageType: Decoder, F[_] : Effect : Logger](queue: Queue[F, Event[MessageType]], subscription: ProjectSubscriptionName, credential: ServiceAccountCredentials, flowControlSettings: Option[FlowControlSettings]): Resource[F, Subscriber] = {
+  def stringSubscriber[F[_]: Effect: StructuredLogger](subscriberConfig: SubscriberConfig, queue: fs2.concurrent.Queue[F, Event[String]]): Resource[F, Subscriber] = {
+    val subscription = ProjectSubscriptionName.of(subscriberConfig.projectTopicName.getProject, subscriberConfig.projectTopicName.getTopic)
+
+    for {
+      credential <- credentialResource(subscriberConfig.pathToCredentialJson)
+      subscriptionAdminClient <- subscriptionAdminClientResource(credential)
+      _ <- createSubscription(subscriberConfig, subscription, subscriptionAdminClient)
+      flowControlSettings = subscriberConfig.flowControlSettingsConfig.map(config =>
+        FlowControlSettings
+          .newBuilder
+          .setMaxOutstandingElementCount(config.maxOutstandingElementCount)
+          .setMaxOutstandingRequestBytes(config.maxOutstandingRequestBytes)
+          .build
+      )
+      sub <- stringSubscriberResource(queue, subscription, credential, flowControlSettings)
+    } yield sub
+  }
+
+  private def subscriberResource[MessageType: Decoder, F[_] : Effect : StructuredLogger](queue: Queue[F, Event[MessageType]], subscription: ProjectSubscriptionName, credential: ServiceAccountCredentials, flowControlSettings: Option[FlowControlSettings]): Resource[F, Subscriber] = {
     val subscriber = for {
       builder <- Sync[F].delay(Subscriber
         .newBuilder(subscription, receiver(queue))
@@ -115,15 +170,29 @@ object GoogleSubscriberInterpreter {
     Resource.make(subscriber)(s => Sync[F].delay(s.stopAsync()))
   }
 
-  private def createSubscription[MessageType: Decoder, F[_] : Effect : Logger](subsriberConfig: SubscriberConfig, subscription: ProjectSubscriptionName, subscriptionAdminClient: SubscriptionAdminClient) = {
+  private def stringSubscriberResource[F[_] : Effect](queue: Queue[F, Event[String]], subscription: ProjectSubscriptionName, credential: ServiceAccountCredentials, flowControlSettings: Option[FlowControlSettings]): Resource[F, Subscriber] = {
+    val subscriber = for {
+      builder <- Sync[F].delay(Subscriber
+        .newBuilder(subscription, stringReceiver(queue))
+        .setCredentialsProvider(FixedCredentialsProvider.create(credential)))
+      builderWithFlowControlSetting <- flowControlSettings.traverse{
+        fcs =>
+          Sync[F].delay(builder.setFlowControlSettings(fcs))
+      }
+    } yield builderWithFlowControlSetting.getOrElse(builder).build()
+
+    Resource.make(subscriber)(s => Sync[F].delay(s.stopAsync()))
+  }
+
+  private def createSubscription[F[_] : Effect : Logger](subsriberConfig: SubscriberConfig, subscription: ProjectSubscriptionName, subscriptionAdminClient: SubscriptionAdminClient): Resource[F, Unit] = {
     Resource.liftF(Async[F].delay(
       subscriptionAdminClient.createSubscription(subscription, subsriberConfig.projectTopicName, PushConfig.getDefaultInstance, subsriberConfig.ackDeadLine.toSeconds.toInt)
     ).void.recover {
-      case _: AlreadyExistsException => ()
+      case _: AlreadyExistsException => Logger[F].info(s"subscription ${subscription} already exists")
     })
   }
 
-  private def subscriptionAdminClientResource[MessageType: Decoder, F[_] : Effect : Logger](credential: ServiceAccountCredentials) = {
+  private def subscriptionAdminClientResource[F[_] : Effect : Logger](credential: ServiceAccountCredentials) = {
     Resource.make[F, SubscriptionAdminClient](Async[F].delay(
       SubscriptionAdminClient.create(
         SubscriptionAdminSettings.newBuilder()
@@ -135,4 +204,7 @@ object GoogleSubscriberInterpreter {
 
 final case class FlowControlSettingsConfig(maxOutstandingElementCount: Long, maxOutstandingRequestBytes: Long)
 final case class SubscriberConfig(pathToCredentialJson: String, projectTopicName: ProjectTopicName, ackDeadLine: FiniteDuration, flowControlSettingsConfig: Option[FlowControlSettingsConfig])
-final case class Event[A](msg: A, consumer: AckReplyConsumer)
+final case class Event[A](msg: A,
+                          traceId: Option[TraceId] = None,
+                          publishedTime: Timestamp,
+                          consumer: AckReplyConsumer)